requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
chromadb>=0.4.22
sentence-transformers>=2.6.0
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.24.0
